backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
google-generativeai==0.8.3
gitpython==3.1.43
numpy==2.1.3
pydantic==2.9.2
pydantic-settings==2.6.1
sse-starlette==2.1.3
httpx==0.27.2
aiofiles==24.1.0
slowapi==0.1.9
